Refactor _pkg_install to handle more than .deb
The previous version was heavily focused on debian install as some of
the applications weren't available in repositories and way out of date.
This new version enable the installation of other mediums like appimage,
this is useful as neovim 0.5.0 dev versions are packaged using appimage
and anyway some distributions don't handle .deb.